src/requirements.txt
Flask==2.0.2
Flask-First==0.9.4
Flask-SQLAlchemy==2.5.1
gunicorn==20.1.0
keras==2.9.0
numpy==1.22.4
Pillow==9.1.1
psycopg2-binary==2.9.3
tensorflow==2.9.1
Werkzeug==2.0.3
